// sealos/frontend/providers/dbprovider/src/types/db.d.ts
export type ReconfigStatusMapType = {
  label: I18nCommonKey;
  value: ReconfigStatus;
  color: string;
};

export interface OpsRequestItemType {
  id: string;
  name: string;
  status: ReconfigStatusMapType;
  startTime: Date;
  namespace: string;
  configurations?: { parameterName: string; newValue: string; oldValue?: string }[];
  switchover?: {
    componentName: string;
    instanceName: string;
  };
}